LayoutTests/imported/w3c/web-platform-tests/workers/name-property-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'message_port.start')

Harness Error (FAIL), message = Unhandled rejection: null is not an object (evaluating 'message_port.start')

PASS name property value for DedicatedWorkerGlobalScope
PASS name property is replaceable for DedicatedWorkerGlobalScope
PASS Declaring name as an accidental global must not cause a harness error for DedicatedWorkerGlobalScope
